requirements.txt
fastapi
uvicorn
pydantic
email-validator
sqlalchemy
python-multipart
pyjwt
dotenv
jinja2